+++ src/services/api/services/postService.ts
@@ -1,5 +1,5 @@
 import { DynamoDB } from "@aws-sdk/client-dynamodb";
-import type { Post } from "../../../shared/types/post";
+import type { Post } from "../types/Post";
 
 export class PostService {
   private db: DynamoDB;
@@ -12,9 +12,9 @@ export class PostService {
 
   async getPost(id: string) {}
 
-  async createPost(post: Post) {}
+  async createPost(post: Omit<Post, "id">) {}
 
-  async updatePost(id: string, post: Post) {}
+  async updatePost(id: string, post: Omit<Post, "id">) {}
 
   async deletePost(id: string) {}
 }
