+++ src/services/api/tsconfig.json
@@ -1,5 +1,1 @@
-{
-    "compilerOptions": {
-        "rootDirs": ["./", "../../shared/types"]
-    }
-}
+{}
